code restructuring and validation middleware
