+++ backend/src/index.ts
@@ -1,7 +1,8 @@
 import cookieParser from "cookie-parser";
+import cors from "cors";
 import dotenv from "dotenv";
 import express from "express";
-import cors from "cors";
+
 import businessRoutes from "./routes/business";
 import categoryRoutes from "./routes/category";
 import userRoutes from "./routes/user";
@@ -11,9 +12,7 @@ const app = express();
 const port = process.env.PORT || 5000;
 
 app.use(express.json());
-
 app.use(cors());
-
 app.use(express.urlencoded({ extended: true }));
 app.use(cookieParser());
